data_strategies: split into multiple files
