+++ lir/data_strategies/__init__.py
@@ -1,0 +1,16 @@
+from lir.data_strategies.labels import CrossValidation, TrainTestSplit
+from lir.data_strategies.pairs import PairsTrainTestSplit
+from lir.data_strategies.predefined import PredefinedCrossValidation, PredefinedTrainTestSplit, RoleAssignment
+from lir.data_strategies.sources import SourcesCrossValidation, SourcesTrainTestSplit
+
+
+__all__ = [
+    'TrainTestSplit',
+    'CrossValidation',
+    'PairsTrainTestSplit',
+    'RoleAssignment',
+    'PredefinedTrainTestSplit',
+    'PredefinedCrossValidation',
+    'SourcesTrainTestSplit',
+    'SourcesCrossValidation',
+]
