+++ tests/data_strategies/test_sources.py
@@ -1,12 +1,10 @@
 import numpy as np
 
-from lir.data.models import FeatureData
-from lir.data_strategies import PairsTrainTestSplit, PredefinedTrainTestSplit, SourcesTrainTestSplit
+from lir.data_strategies import SourcesTrainTestSplit
 from lir.datasets.synthesized_normal_multiclass import (
     SynthesizedDimension,
     SynthesizedNormalMulticlassData,
 )
-from lir.transform.pairing import InstancePairing
 
 
 def test_multiclass_train_test_split():
@@ -45,30 +43,3 @@ def test_multiclass_train_test_split_seed():
     for data_train, data_test in strategy.apply(data):
         assert data_train != ref_train
         assert data_test != ref_test
-
-
-def test_paired_train_test_split():
-    instances = FeatureData(features=np.ones(20), source_ids=np.arange(10).repeat(2))
-    pairs = InstancePairing().pair(instances)
-    training_pairs, test_pairs = next(iter(PairsTrainTestSplit(test_size=0.5, seed=0).apply(pairs)))
-    assert len(training_pairs) == len(test_pairs) == 10 * 9 / 2
-    assert len(np.unique(training_pairs.source_ids)) == 5, 'half of the sources are for training'
-    assert len(np.unique(test_pairs.source_ids)) == 5, 'half of the sources are for testing'
-    assert len(np.unique(np.concatenate([test_pairs.source_ids, training_pairs.source_ids]))) == 10, (
-        'all sources are for training or testing'
-    )
-
-
-def test_predefined_role_splitter():
-    instances = FeatureData(
-        features=np.arange(16).reshape(-1, 2), role_assignments=np.array(['train', 'test']).repeat(4)
-    )
-    splitter = PredefinedTrainTestSplit()
-    splits = list(splitter.apply(instances))
-    assert len(splits) == 1, 'exactly one split'
-
-    training_set, test_set = splits[0]
-    expected_training_features = np.arange(8).reshape(-1, 2)
-    expected_test_features = expected_training_features + 8
-    assert training_set == FeatureData(features=expected_training_features, role_assignments=np.array(['train'] * 4))
-    assert test_set == FeatureData(features=expected_test_features, role_assignments=np.array(['test'] * 4))
